refactor(medications): SAV-1141: extract shared label print preview component
Create MedicationLabelPrintPreview component to eliminate duplication
across DispenseMedicationWorkflowModal, EditMedicationDispenseModal,
and MedicationLabelPrintModal. This ensures all three screens use
identical print styles and layout, guaranteeing consistent centering
and rendering behaviour.
