+++ packages/web/app/components/PatientPrinting/index.js
@@ -7,6 +7,7 @@ export * from './modals/CovidVaccineCertificateModal';
 export * from './modals/VaccineCertificateModal';
 export * from './modals/DeathCertificateModal';
 export * from './modals/MedicationLabelPrintModal';
+export { MedicationLabelPrintPreview } from './printouts/MedicationLabelPrintPreview';
 
 export { PrintLetterhead } from './printouts/reusable/PrintLetterhead';
 export { PrintPortal } from './PrintPortal';
